Support for Libertine sans-serif font added. Extra check for Garamond font
Libertine sans-serif font added to list of supported fonts. Libertine
can be enabled by adding "libertine" to the documentclass options
within the LaTeX document.

Mathdesign package now classifies Garamond as a non-free font and is
not included in standard release. A check for the necessary package is
made before attempting to change to Garamond.